app/requirements.txt
asgiref==3.2.10
Babel==2.8.0
certifi==2020.6.20
chardet==3.0.4
Django==3.1
django-phone-verify==2.0.0
django-phonenumber-field==4.0.0
djangorestframework==3.11.1
idna==2.10
mysqlclient==2.0.1
nexmo==2.4.0
phonenumbers==8.12.7
PyJWT==1.7.1
python-dotenv==0.14.0
pytz==2020.1
python-dotenv==0.14.0
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
twilio==6.44.2
urllib3==1.25.10
